==> code/step_01c_create_opm_nondod_binaries_2014_2016.ipynb <==
from pathlib import Path
import pandas as pd
import json
from pandas.api.types import union_categoricals
from itertools import islice

opm_status_path_dict = {}

# Create dict of file paths
opm_status_path = Path('../raw_data/opm-federal-employment-data/data/2014-09-to-2016-09/non-dod/status')
for file_path in opm_status_path.iterdir():
    file_name = file_path.name

    year = file_name[8:12]
    qtr = file_name[12:14]
    date_str = year + '-' + qtr

    file_date = pd.to_datetime(date_str)
    year = str(file_date.year)
    qtr = str(file_date.quarter)
    opm_status_path_dict.setdefault(year, {}) # Create empty dict for that year if not yet created
    opm_status_path_dict[year][qtr] = str(file_path)

opm_status_path = Path('../raw_data/opm-federal-employment-data/data/2016-12-to-2017-03/non-dod/status')
for file_path in opm_status_path.iterdir():
    file_name = file_path.name
    year = file_name[24:28]
    qtr = file_name[28:30]
    date_str = year + '-' + qtr

    file_date = pd.to_datetime(date_str)
    year = str(file_date.year)
    qtr = str(file_date.quarter)
    opm_status_path_dict.setdefault(year, {}) # Create empty dict for that year if not yet created
    opm_status_path_dict[year][qtr] = str(file_path)

file_path = opm_status_path_dict['2014']['3']

df = pd.read_csv(file_path, delimiter = '|', header = None)

df
