==> code/step_02a_clean_opm_nondod_binaries.ipynb <==
from pathlib import Path
import pandas as pd
import json
from pandas.api.types import union_categoricals
from itertools import islice

# Load dicts of paths of binaries
with open('../code_output/opm_nondod_dynamic_pre2014_feather_path_dict.json', 'r') as infile:
    opm_dynamic_feather_path_dict = json.load(infile)

with open('../code_output/opm_nondod_status_pre2014_feather_path_dict.json', 'r') as infile:
    opm_status_feather_path_dict = json.load(infile)

# Load list of columns that are categorical
with open('../code_output/opm_dynamic_cat_col_list.json') as infile:
    opm_dynamic_cat_col_list = json.load(infile)

with open('../code_output/opm_status_cat_col_list.json') as infile:
    opm_status_cat_col_list = json.load(infile)

# Store categorical variable map for each categorical variable and year in nested dict
opm_dynamic_cat_index_dict = {}
opm_status_cat_index_dict = {}

for col in opm_dynamic_cat_col_list:
    opm_dynamic_cat_index_dict[col] = {}
    for year, year_dict in opm_dynamic_feather_path_dict.items():
        opm_dynamic_cat_index_dict[col][year] = {}

for col in opm_status_cat_col_list:
    opm_status_cat_index_dict[col] = {}
    for year, year_dict in opm_status_feather_path_dict.items():
        opm_status_cat_index_dict[col][year] = {}

for year, qtr_dict in opm_dynamic_feather_path_dict.items():
    for qtr, str_path in qtr_dict.items():
        df = pd.read_feather(Path(str_path))

        for col in opm_dynamic_cat_col_list:
            cat_index = pd.CategoricalIndex(df[col])
            opm_dynamic_cat_index_dict[col][year][qtr] = cat_index

for year, qtr_dict in opm_status_feather_path_dict.items():
    for qtr, str_path in qtr_dict.items():
        df = pd.read_feather(Path(str_path))

        for col in opm_status_cat_col_list:
            cat_index = pd.CategoricalIndex(df[col])
            opm_status_cat_index_dict[col][year][qtr] = cat_index

# Find the union across years of each categorical variable, and store resulting union in dict
opm_dynamic_uc_dict = {}
for col, year_dict in opm_dynamic_cat_index_dict.items():
    cat_index_list = []
    for year, cat_index_dict in year_dict.items():
        for qtr, cat_index in cat_index_dict.items():
            cat_index_list.append(cat_index)
    
    opm_dynamic_uc_dict[col] = union_categoricals(cat_index_list)

opm_status_uc_dict = {}
for col, year_dict in opm_status_cat_index_dict.items():
    cat_index_list = []
    for year, cat_index_dict in year_dict.items():
        for qtr, cat_index in cat_index_dict.items():
            cat_index_list.append(cat_index)
    
    opm_status_uc_dict[col] = union_categoricals(cat_index_list)

# Update categorical mapping for all binaries, to enable type retention when concatenating
for year, qtr_dict in opm_dynamic_feather_path_dict.items():
    for qtr, str_path in qtr_dict.items():
        df = pd.read_feather(Path(str_path))

        for col, uc in opm_dynamic_uc_dict.items():
            df[col] = pd.Categorical(df[col], categories = uc.categories)

        df.to_feather(Path(str_path))

for year, qtr_dict in opm_status_feather_path_dict.items():
    for qtr, str_path in qtr_dict.items():
        df = pd.read_feather(Path(str_path))

        for col, uc in opm_status_uc_dict.items():
            df[col] = pd.Categorical(df[col], categories = uc.categories)

        df.to_feather(Path(str_path))

# Print out all categorical values
for col, cats in opm_dynamic_uc_dict.items():
    col = col.replace('/', '-')
    with open(f'../code_output/opm_nondod_dynamic_pre2014_{col}_cats.txt', 'w') as outfile:
        list = cats.categories.values.tolist()
        list.sort()
        for item in list:
            outfile.write(item + '\n')

# Print out all categorical values
for col, cats in opm_status_uc_dict.items():
    col = col.replace('/', '-')
    with open(f'../code_output/opm_nondod_status_pre2014_{col}_cats.txt', 'w') as outfile:
        list = cats.categories.values.tolist()
        list.sort()
        for item in list:
            outfile.write(item + '\n')
