==> code/step_06a_analysis.ipynb <==
from pathlib import Path
import pandas as pd
import json
from itertools import islice
import re

# Save dictionary for paths of binaries
with open('../code_output/oews_opm_merged_feather_path_dict.json') as infile:
    df_path_dict = json.load(infile)

combined_df = pd.DataFrame()
for year, df_path in df_path_dict.items():
    df = pd.read_feather(Path(df_path))
    df['year'] = year

    combined_df = pd.concat([combined_df, df])

combined_df.to_stata('../stata/opm_oews_merged.dta')
